Reestructuración de directorios del proyecto
Las buenas practicas de Go indica como se debe dividir un proyecto:
https://github.com/golang-standards/project-layout
