+++ internal/projectpath/projectpath.go
@@ -1,0 +1,13 @@
+package projectpath
+
+import (
+	"path/filepath"
+	"runtime"
+)
+
+var (
+	_, b, _, _ = runtime.Caller(0)
+
+	// Root folder of this project
+	Root = filepath.Join(filepath.Dir(b), "../..")
+)
